requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
pillow==10.3.0
opencv-python==4.9.0.80
pytesseract==0.3.13
nltk==3.8.1
python-dotenv==1.0.1
web3==6.15.1
pandas==2.2.2
numpy==1.26.4
pdf2image==1.17.0
pydantic==2.7.1
setuptools>=68.2.2 
wheel>=0.42.0